moving utility functions out to the utils folder
