+++ src/realtime_api_async_python/test_api_key.py
@@ -9,7 +9,7 @@ openai.api_key = os.getenv("OPENAI_API_KEY")
 
 try:
     response = openai.chat.completions.create(
-        model="gpt-3.5-turbo", messages=[{"role": "user", "content": "Hello"}]
+        model="gpt-4o-mini", messages=[{"role": "user", "content": "Hello"}]
     )
     print("API Key is valid. Response: ", response.choices[0].message.content)
 except Exception as e:
