+++ src/realtime_api_async_python/utils/encode_audio.py
@@ -1,0 +1,4 @@
+import base64
+
+def encode_audio(audio_bytes):
+    return base64.b64encode(audio_bytes).decode("utf-8")
