+++ src/realtime_api_async_python/utils/logging.py
@@ -1,0 +1,67 @@
+import os
+import logging
+import json
+from datetime import datetime
+from dotenv import load_dotenv
+
+load_dotenv()
+
+# Set up logging
+logging.basicConfig(
+    level=logging.INFO,
+    format="%(asctime)s.%(msecs)03d - %(levelname)s - %(message)s",
+    datefmt="%H:%M:%S",
+)
+
+# Function to log WebSocket events
+def log_ws_event(direction, event):
+    event_type = event.get("type", "Unknown")
+    event_emojis = {
+        "session.update": "🛠️",
+        "session.created": "🔌",
+        "session.updated": "🔄",
+        "input_audio_buffer.append": "🎤",
+        "input_audio_buffer.commit": "✅",
+        "input_audio_buffer.speech_started": "🗣️",
+        "input_audio_buffer.speech_stopped": "🤫",
+        "input_audio_buffer.cleared": "🧹",
+        "input_audio_buffer.committed": "📨",
+        "conversation.item.create": "📥",
+        "conversation.item.delete": "🗑️",
+        "conversation.item.truncate": "✂️",
+        "conversation.item.created": "📤",
+        "conversation.item.deleted": "🗑️",
+        "conversation.item.truncated": "✂️",
+        "response.create": "➡️",
+        "response.created": "📝",
+        "response.output_item.added": "➕",
+        "response.output_item.done": "✅",
+        "response.text.delta": "✍️",
+        "response.text.done": "📝",
+        "response.audio.delta": "🔊",
+        "response.audio.done": "🔇",
+        "response.done": "✔️",
+        "response.cancel": "⛔",
+        "response.function_call_arguments.delta": "📥",
+        "response.function_call_arguments.done": "📥",
+        "rate_limits.updated": "⏳",
+        "error": "❌",
+        "conversation.item.input_audio_transcription.completed": "📝",
+        "conversation.item.input_audio_transcription.failed": "⚠️",
+    }
+    emoji = event_emojis.get(event_type, "❓")
+    icon = "⬆️ - Out" if direction == "outgoing" else "⬇️ - In"
+    logging.info(f"{emoji} {icon} {event_type}")
+
+def log_runtime(function_or_name: str, duration: float):
+    jsonl_file = os.getenv("RUN_TIME_TABLE_LOG_JSON", "runtime_time_table.jsonl")
+    time_record = {
+        "timestamp": datetime.now().isoformat(),
+        "function": function_or_name,
+        "duration": f"{duration:.4f}",
+    }
+    with open(jsonl_file, "a") as file:
+        json.dump(time_record, file)
+        file.write("\n")
+
+    logging.info(f"⏰ {function_or_name}() took {duration:.4f} seconds")
