+++ src/realtime_api_async_python/utils/timeit_decorator.py
@@ -1,0 +1,61 @@
+import os
+import asyncio
+import functools
+import time
+from datetime import datetime
+import json
+from dotenv import load_dotenv
+
+RUN_TIME_TABLE_LOG_JSON = os.getenv("RUN_TIME_TABLE_LOG_JSON","runtime_time_table.jsonl")
+
+def timeit_decorator(func):
+    @functools.wraps(func)
+    async def async_wrapper(*args, **kwargs):
+        start_time = time.perf_counter()
+        result = await func(*args, **kwargs)
+        end_time = time.perf_counter()
+        duration = round(end_time - start_time, 4)
+        print(f"⏰ {func.__name__}() took {duration:.4f} seconds")
+
+        jsonl_file = RUN_TIME_TABLE_LOG_JSON
+
+        # Create new time record
+        time_record = {
+            "timestamp": datetime.now().isoformat(),
+            "function": func.__name__,
+            "duration": f"{duration:.4f}",
+        }
+
+        # Append the new record to the JSONL file
+        with open(jsonl_file, "a") as file:
+            json.dump(time_record, file)
+            file.write("\n")
+
+        return result
+
+    @functools.wraps(func)
+    def sync_wrapper(*args, **kwargs):
+        start_time = time.perf_counter()
+        result = func(*args, **kwargs)
+        end_time = time.perf_counter()
+        duration = round(end_time - start_time, 4)
+        print(f"⏰ {func.__name__}() took {duration:.4f} seconds")
+
+        jsonl_file = RUN_TIME_TABLE_LOG_JSON
+
+        # Create new time record
+        time_record = {
+            "timestamp": datetime.now().isoformat(),
+            "function": func.__name__,
+            "duration": f"{duration:.4f}",
+        }
+
+        # Append the new record to the JSONL file
+        with open(jsonl_file, "a") as file:
+            json.dump(time_record, file)
+            file.write("\n")
+
+        return result
+
+    return async_wrapper if asyncio.iscoroutinefunction(func) else sync_wrapper
+
